Crate nvimpam

The nvimpam binary. Needs to be connected to neovim by stdin/stdout like this (assuming it is in your PATH):

let s:bin = 'nvimpam'
let s:id = jobstart([s:bin], { 'rpc': v:true }

It will automatically notify neovim of its activity, request the whole buffer and parse it for folds. After that, it sends the folds to neovim.

As a performance optimization, pass the filename as a first argument:

let s:bin = 'nvimpam'
let s:file = expand('%:p')
let s:id = jobstart([s:bin.' '.s:file], { 'rpc': v:true }

Nvimpam will load the file from disc instead of requesting it over RPC, which is quite a bit faster (mostly probably because the file is cached by your OS since it was loaded by neovim just before).

If you want logging, set the following environment variables:

NVIMPAM_LOG_FILE is the path to the log file (no logging if this is empty)

NVIMPAM_LOG_LEVEL can be one of error, warn, info, debug and trace, in ascending order of verbosity. The default is warn.